# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitrack",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exercise-log/{user_id}": {
            "get": {
                "tags": [
                    "ExerciseLog"
                ],
                "summary": "Obtener logs de ejercicio por ID de usuario",
                "description": "Devuelve todos los registros de ejercicio de un usuario específico.",
                "operationId": "getExerciseLogsByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de logs de ejercicio del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "userId": {
                                                "type": "integer",
                                                "example": 123
                                            },
                                            "exerciseId": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-06-01"
                                            },
                                            "weight": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 50.5
                                            },
                                            "reps": {
                                                "type": "integer",
                                                "example": 10
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ExerciseLog"
                ],
                "summary": "Insertar un nuevo log de ejercicio para un usuario",
                "description": "Inserta un nuevo registro de ejercicio para un usuario si no existe ya con la misma fecha y ejercicio.",
                "operationId": "insertExerciseLogByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "exercise_id",
                                    "date",
                                    "weight",
                                    "reps"
                                ],
                                "properties": {
                                    "exercise_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-04"
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75
                                    },
                                    "reps": {
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Log insertado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercise Log insertado exitosamente"
                                        },
                                        "exercise_log": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "El log ya existe para ese usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercise Log ya existe para ese usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un error inesperado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exercise-log/{user_id}/{exercise_Id}": {
            "delete": {
                "tags": [
                    "ExerciseLog"
                ],
                "summary": "Eliminar un log de ejercicio por ID de usuario e ID de ejercicio",
                "description": "Elimina un registro de ejercicio específico de un usuario.",
                "operationId": "deleteExerciseLogByUserIdAndExerciseId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "exercise_Id",
                        "in": "path",
                        "description": "ID del log de ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercise Log deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Exercise Log not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token en formato Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}